refactor: move logic to a function
