+++ frappe/boot.py
@@ -528,16 +528,25 @@ def get_sentry_dsn():
 
 
 def get_sidebar_items():
+	from frappe.utils.install import auto_generate_sidebar_from_module
+
 	sidebars = frappe.get_all(
 		"Workspace Sidebar", fields=["name", "header_icon"], filters={"name": ["not like", "%My Workspaces%"]}
 	)
+	module_sidebars = auto_generate_sidebar_from_module()
+	sidebars.extend(module_sidebars)
 	add_user_specific_sidebar(sidebars)
 	sidebar_items = {}
 
 	for s in sidebars:
-		w = frappe.get_doc("Workspace Sidebar", s.get("name"))
-		sidebar_items[s["name"].lower()] = {
-			"label": s.get("name"),
+		sidebar_title = s.get("name")
+		if sidebar_title:
+			w = frappe.get_doc("Workspace Sidebar", sidebar_title)
+		else:
+			sidebar_title = s.title
+			w = s
+		sidebar_items[sidebar_title.lower()] = {
+			"label": sidebar_title,
 			"items": [],
 			"header_icon": s.get("header_icon"),
 			"module": w.module,
@@ -569,11 +578,11 @@ def get_sidebar_items():
 					"ref_doctype": ref_doctype,
 				}
 			if (
-				"My Workspaces" in s["name"]
+				"My Workspaces" in sidebar_title
 				or si.type == "Section Break"
 				or w.is_item_allowed(si.link_to, si.link_type)
 			):
-				sidebar_items[s.get("name").lower()]["items"].append(workspace_sidebar)
+				sidebar_items[sidebar_title.lower()]["items"].append(workspace_sidebar)
 
 	old_name = f"my workspaces-{frappe.session.user.lower()}"
 	if old_name in sidebar_items.keys():
