+++ frappe/patches.txt
@@ -252,4 +252,3 @@ frappe.patches.v16_0.auto_generate_desktop_icon_and_sidebar
 frappe.patches.v16_0.add_private_workspaces_to_sidebar
 frappe.core.doctype.communication_link.patches.copy_communication_date_to_link
 frappe.core.doctype.communication.patches.drop_ref_dt_dn_index
-frappe.patches.v16_0.auto_generate_sidebar_from_modules
